Create custom SimpleZlib Implementation
The absolute morons at BSG compiled a custom version of ComponentAce's Zlib library exclusively to add a helper class. With the help of DotPeek I was able to create a (arguably better) implementation of SimpleZlib that we can use. This currently doesn't use pooling though but I imagine that is overkill for the manager