Fixed Server stacking and deadlocking, refactored code, translated Install Page.
Fixed the stacking of the page by also checking if the server is running in Play Page View Model.

Deadlocking was causing by the unhandled exception which user never get when he was closing the server (Probably because multiple servers runned at the same time in launcher).

Translated and updated missing keys in the in Install Page. Fixed not added "File-Service" translation.

Added ability to clear the console output if user wants it with additional check asking user if he wants to remove entire console output.

Fixed crocked visuals of the Install Page. Server Page.

Refactored few functions to make translation work, or fixed spelling mistake in one of the functions.

Removed useless http-client dependency. That was assigned, but never actually used.