Create DllManager framework for DLL overrides
Some DLL managers are not implemented, one of which is DxvkNvapiManager.cs because I cannot test it (I have an AMD GPU, therefore NVAPI does nothing for me).
Default config has changed as I have done some testing and (at least on my machine) only DXVK override needs to be enabled.

Custom DLL overrides will be supported soon. I just have to figure out how to implement the front-end for it.
I also moved the existing wine configuration over to LinuxConfig.cs
I have made changes in the localization as well, but those changes will be added in the next commit together with the changes to the LinuxSettingsPage